requirements.txt
# Python dependencies for AI Search Clone

flask
openai
python-dotenv
requests

# Using uv for installation, but listing dependencies here 